Remove manifest demo examples and refactor manifest logic
- Delete publish_with_manifest.rs and store_manifest_demo.rs examples -
Move URL matching and extension indexing from StoreManifest to
LocalStoreManifest - Update LocalStore and Store trait to support
URL/domain extension queries - Refactor StoreManager to use new URL
matching methods - Clean up StoreManifest struct and related tests
